Moved GPO notifications into ProfileManager class to avoid needing a listener for each profile
