+++ DPCService/Models/GPOEventArgs.cs
@@ -1,0 +1,8 @@
+﻿using System;
+
+namespace DPCService.Models
+{
+    public class GPOEventArgs : EventArgs
+    {
+    }
+}
